initial commit, added media and changed structure
we're adding media and modifying,as an initial stage of the revamping of the
site. The box elements are already in place with working animation.

changes to be done now should be focused on getting thing to work
aesthetically, among them the following:
+ completing the list of "skills" (if you can call them that)
+ agreeing on the animations for the boxes
+ coming up with ideas for the third section.

branching has been suggested for testign the different ideas. We'll see
about that.