massive refactoring, all components in own folders
The original project had many components, all of them in a single 'App'
file, or a 'utils' file (for miscellaneous comps). This commit aimed at
placing them in a hierarchy of folders, mirroring the structure of the
project itself.

Several other cleanup changes were made. The project won't run for now.