moved to styled-components, rewrote navbar
in this commit, all inline styles were redone using the library
styled-components, which makes some problems easier, such as
using CSS pseudo-selectors, something tht seems impossible to do
with the inline syntax.

Additionally, the navbar component got redone from scratch. this
was entirely in React/styled, so we do not need the react-boostrap
library anymore. We can do it ourselves.