Rewired animation components
in this commit, i defined the architecture of the components
related to the animation.

there will be an inner part taking care of the sprite movement,
while an external container tracks the component across the
board. The component is rendered via render props and sent
from the external component to the inner one.