implemented first version of path functions
In this commit, a significant effort was put forward to write
the necessary logic to simulate character animations.

There are two. "walking" handles simple sideways movement. "falling"
simulates gravity as an object is dropped in midair. The functions are
only barely implemented and they will have to be fine-tuned to account
for every pixel and avoid clipping.

Additional changes as some composition problems were fixed with the
dependencies, and a few other files go renamed.