Made site functional, added cube animation
In this commit, the website is again functional, albeit some buttons
still crash the app.

A 3D animation with a rotating cube has been added. It looks pretty
neat.

In the next few commits, I should tidy up the layout and presentation
of the third section of the webpage. Interactive animations are for
now not included. Once everything has fallen into place, we'll begin
to add the animations and state changes.