Adjusted size of cube, fixed video
In this relatively small commit, I changes the size of the cube
to make sure it didn't eat up parts of the site.

At the same time, The background video looks a lot better, thanks
to a second instance of the same video playing in the background
with a slight blur. This seems to have had a bad effect on
performance. So tomorrow, we'll write a function that pause()s and
load()s the two vids as soon as window.scrollY reaches a certain
threshold value.

also everything else. (help)