Added tests, improved presentation
In this commit, the first tests haven been added, concerning two of the
components of the website: Catalog and Navbar. The tests are thus far
incomplete, however, tomorrow, I will finish a tutorial and write
the remaining tests of the components.

Using the test driven development approach, the next commit will
include the fixed Catalog function, as I write the remaining tests.