Fixed cube animation, rewrote tests for Catalog
In this commit, we finally get the cube right and the buttons now
are capable of switching the skill shown in the Catalog to display
the different abilities.

In the same page, new tests were written for the Catalog, including
one for the function that manages the change in the displayed skill.
These tests are the fir of many that will be written to guarantee
the function of our app.