Finished cube anim, added social media buttons
in this commit, the cube animation is completed. Still looks a bit
jerky, but that can be fixed in future iterations. Now, when pressing
the buttons, the cube will shrink and spin around, while displaying
the logo of each technology.

The other thing was the addition of social media buttons at the bottom.
they have some basic styling, with on hover effects, but still need some
work. Improving how they fit in the background would be a start.

Finally, some additional tests have been added.