innerHTML in TextBoxes now working, bugfixes
This commit was aimed mainly at getting text right.

In the Text-Boxes, a bit of text decoration was added. In order for it
to show without being escaped, it had to be set as innerHTML. This
allows for formatting to show up.

Additionally, some other changes were made. The name for spanish was
updated to "español".